Convert logo from png to webp and move css files to styles folder
