+++ client/public/index.html
@@ -30,7 +30,6 @@
     -->
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Press+Start+2P&text=1234567890ABCDEFGHIJKLMNOPQRSTUVÄÖWXYZabcdefghijklmnopqrstuvwxyzäö%20%22%27" rel="stylesheet">
     <title>MapDasher | Geography map game</title>
   </head>
   <body>
