+++ client/src/assetts/styles/index.css
@@ -16,6 +16,13 @@ code {
     monospace;
 }
 
+@font-face {
+  font-family: 'Press Start 2P';
+  font-style: normal;
+  font-weight: 400;
+  src: url(https://fonts.gstatic.com/l/font?kit=e3t4euO8T-267oIAQAu6jDQyK0nXmIZswKAFNKcu2V9vRSO02AVwAQsEqUK7PQtlKU6jKz3AplrjZh-1HAm32rJX6-_Qu6N_SnhyWXCj-VNrB_ti4ZdY5Xabsb5xdP_XPRg56uLqiySF9drJRnX5NC2oPxN1SUOLrtOQW5dzltm6P7L3w1XJ&skey=24cc298767068e24&v=v15) format('woff2');
+}
+
 h1, .game-over { font-family: 'Press Start 2P', cursive; }
 
 /* leaflet overrides */
