+++ client/src/components/Elements/MapDasherLogo/MapDasherLogo.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import MapDasherImg from '../../../assetts/images/mapdasher.png';
+import MapDasherImg from '../../../assetts/images/mapdasher.webp';
 
 const MapDasherLogo = () => {
   return (
